+++ tests/test_code.py
@@ -63,6 +63,17 @@ class TestPythonEvaluator:
         assert evaluator._evaluate_code('event.data[\'a\'] == 1', additional_context={'event': Event('test', a=1)})
         assert evaluator._evaluate_code('event.name == \'test\'', additional_context={'event': Event('test')})
 
+    def test_setdefault(self, evaluator, mocker):
+        statechart = mocker.MagicMock()
+        statechart.preamble = 'setdefault("x", 2)'
+        evaluator.execute_statechart(statechart)
+        assert evaluator.context['x'] == 1
+
+        statechart.preamble = 'b = setdefault("a", 0)'
+        evaluator.execute_statechart(statechart)
+        assert evaluator.context['a'] == 0
+        assert evaluator.context['b'] == 0
+
     def test_execution(self, evaluator):
         evaluator._execute_code('a = 1')
         assert evaluator.context['a'] == 1